+++ include/kaguya/scene/Aggregation.h
@@ -14,7 +14,7 @@ namespace kaguya {
 
         class Aggregation : public Intersectable {
         public:
-            virtual bool intersect(Ray &ray, SurfaceInteraction &si, double minStep, double maxStep) const override;
+            virtual bool intersect(Ray &ray, SurfaceInteraction &si, Float minStep, Float maxStep) const override;
 
             virtual const std::vector<std::shared_ptr<Intersectable>> aggregation() const = 0;
